refactor: delete the unrendered EnhancedCoTMarker store; cotEvents is the single source
TAKService published two marker stores: cotEvents (rendered by both map
engines via MapViewController.cotMarkers) and enhancedMarkers (a richer
UID-keyed dict with position history that NO view rendered).
CoTEventHandler wrote both on every position update; writers that
touched only enhancedMarkers (the old Meshtastic publish path) silently
never appeared on the map, and any future divergence desynced the
stores. enhancedMarkers' only consumers were BloodhoundService (which
also subscribes to cotEvents and builds its own position history) and
the never-assigned onMarkerUpdated callback.

Deleted:
- TAKService.enhancedMarkers + updateEnhancedMarker/getMarker/
  getAllMarkers/removeStaleMarkers/shouldAddToHistory + history config
  + onMarkerUpdated
- CoTEventHandler's dual writes (position + waypoint handlers) and the
  removeEvent dual delete
- BloodhoundService's $enhancedMarkers subscription,
  processEnhancedMarkers, and the EnhancedCoTMarker updateTrack
  overload (the cotEvents subscription already feeds the same tracks)
- EnhancedCoTMarker.swift (type + UnitAffiliation/UnitType/CoTPosition,
  all unused elsewhere), plus the dead UI that depended on it:
  CustomMarkerAnnotation.swift, MarkerInfoPanel's marker panel

Live shared helpers that were co-located with the dead UI (InfoRow,
cornerRadius(_:corners:)/RoundedCorner, Color.uiColor) moved to
Shared/UI/Components/InfoRow.swift — they're used by
EmergencyBeaconView, DigitalPointerView, ConnectionStatusWidget,
RoutePlanningView, and MapViewController.

MeshtasticManager.clearMeshMarkersFromMap now actually works: it
removes mesh markers via CoTEventHandler.removeEvent instead of
printing that markers 'will expire' against a store the map never read.
